refactor(level): 将关卡运行时相关代码挪回关卡相关目录
