+++ OpenSolarMax.Game/Screens/Pages/ChapterPage.cs
@@ -1,7 +1,6 @@
 using Microsoft.Xna.Framework.Graphics;
 using OpenSolarMax.Game.Level;
 using OpenSolarMax.Game.Modding;
-using OpenSolarMax.Game.Screens.Models;
 using OpenSolarMax.Game.Screens.ViewModels;
 using OpenSolarMax.Game.Screens.Views;
