+++ OpenSolarMax.Game/Screens/ViewModels/LevelsViewModel.cs
@@ -6,7 +6,6 @@ using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using OpenSolarMax.Game.Level;
 using OpenSolarMax.Game.Modding;
-using OpenSolarMax.Game.Screens.Models;
 using OpenSolarMax.Game.Screens.Pages;
 using OpenSolarMax.Game.Screens.Transitions;
 using OpenSolarMax.Game.UI;
